Improved resize handling with aspect-locking
Before, the image window quickly behaved weirdly when repeatedly
resizing the window with active aspect-locking. Move the mouse
a few times up and down, and the window becomes very small...

This was due to considering horizontal and vertical movement of
the mouse, it appears.
We now only consider the vertical movement and it works much nicer
IMO, even though only vertical mouse movement resizes the window.